requirements.txt
substrate_interface==1.6.0
substrateinterface==1.0.0
uagents==0.3.1
paho-mqtt==1.6.1
